+++ include/nc_runtime/nc_vector.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "nc_runtime/nc_types.h"
+
+struct NcVector2
+{
+    float x, y;
+};
+
+static forceinline NcVector2 vec2(float x, float y)
+{
+    return NcVector2{x, y};
+}