+++ test/stdafx_nc_runtime_test.h
@@ -3,5 +3,6 @@
 #include "gtest/gtest.h"
 
 #include "nc_runtime/nc_types.h"
+#include "nc_runtime/nc_string.h"
 
 using namespace std;